Split the notebook-path PPO loop out of jax_trainer into jax_train_fn
Phase K (cleanup), entry ST6 from docs/reviews/RL_PIPELINE_GAP_REVIEW_2026_08.md
§6.4. jax_trainer.py held two parallel PPO training loops: the module-level
train() + _build_jit_fns (with TrainConfig / TrainResult) that only the JAX
notebook and its tests use, and the JaxTrainer class the CLI uses. The
functional loop now lives in environments/shared/jax_train_fn.py; jax_trainer
re-exports the four public names (declared in __all__, which also resolves
the F811 that a bare re-export of `train` beside JaxTrainer.train would
raise), so the notebook's import is unchanged. The two helpers both loops
call since Phase J (_bootstrap_truncations, _build_batched_value) move to
jax_trainer_types, the module both already import, so neither loop depends
on the other. Every code line and docstring moved verbatim (checked as a
line multiset); jax_trainer.py goes from 1562 to 624 lines. Test
monkeypatch targets follow the names to where they are now looked up.
